Add formula tweaks for native go, rust, openjdk
Make the fairly simple tweaks required to get these three languages running on M1

Golang’s main tweak is using the v1.16 preview to do bootstrapping, as this is both compiled for M1 and capable of compiling for M1

Rust’s only change is conditionally using the aarch64 darwin minimal environment for bootstrapping, as it’s got less cross-compilation quirks than trying to use the x86 package through rosetta

Openjdk is a mess. I won’t claim to understand half of what’s going on here, but it seems to work alright I think. Occasionally the URL needs updated from jdk.java.net but the main use for java these days seems to be as a build dependency anyway, so… It works well enough for that, it would seem